+++ memberships/views.py
@@ -4,7 +4,6 @@ from django.urls import reverse
 from django.views.decorators.csrf import csrf_exempt
 from django.conf import settings
 from django.contrib.auth import authenticate, login, logout
-from django.contrib.auth.models import Permission, User
 from urllib.parse import parse_qs, urlparse
 
 import urllib.request
@@ -15,8 +14,9 @@ from django.contrib.auth.decorators import login_required
 from django.utils.decorators import method_decorator
 
 from funky_time import epoch_to_datetime
+from payments import handle_stripe_payment
 from .forms import *
-from .models import Member, Membership, FailedPayment, Payment
+from .models import Member, Membership
 from .services import StripeGateway
 
 
@@ -181,42 +181,11 @@ def stripe_webhook(request):
             json.loads(request.body), settings.STRIPE_SECRET_KEY
         )
         stripe_webhook = StripeWebhook()
-        if event["type"] == "invoice.payment_failed":
-            f_payment = FailedPayment.objects.create(
-                stripe_customer_id=event["data"]["object"]["customer"],
-                stripe_subscription_id=event["data"]["object"]["subscription"],
-                stripe_event_type=event["type"],
-            )
-        if event["type"] == "invoice.payment_succeeded":
-            member = Member.objects.get(email=event["data"]["object"]["customer_email"])
-            log_payment(event, member)
-            update_last_payment(event, member)
-            add_user_sand_permission(member)
+        handle_stripe_payment(event)
 
         return stripe_webhook.handle(event)
     except ValueError as e:
         return HttpResponse("Failed to parse stripe payload", status=400)
-
-
-def log_payment(event, member):
-    payment = Payment.objects.create(
-        member=member,
-        stripe_subscription_id=event["data"]["object"]["subscription"],
-    )
-
-
-def update_last_payment(event, member):
-    # Store payment DateTime in membership model
-    membership = Membership.objects.get(member=member)
-    membership.last_payment_time = epoch_to_datetime(event["created"])
-    membership.save()
-
-
-def add_user_sand_permission(member):
-    # Give user 'has_sand_membership' permission
-    user = User.objects.get(id=member.user_id)
-    perm = Permission.objects.get(codename="has_sand_membership")
-    user.user_permissions.add(perm)
 
 
 @login_required()
